requirements.txt
Django
psycopg2
djangorestframework
django-grappelli
